KP-14563 split audit log on server

+++ src/Core/SharedKernels/Enumerator/Enumerator/Implementation/Services/WorkspaceService.cs
@@ -32,7 +32,8 @@ namespace WB.Core.SharedKernels.Enumerator.Implementation.Services
                 foreach (var removedWorkspace in removedWorkspaces)
                 {
                     var workspaceDirectory = fileSystemAccessor.CombinePath(settings.PathToDatabaseDirectory, removedWorkspace.Name);
-                    fileSystemAccessor.DeleteDirectory(workspaceDirectory);
+                    if (fileSystemAccessor.IsDirectoryExists(workspaceDirectory))
+                        fileSystemAccessor.DeleteDirectory(workspaceDirectory);
                 }
                 workspaceRepository.Remove(removedWorkspaces);
             }
